Refactor method documentation logic
Removed method related documentation logic to a seperate file and flattened
the method related Documenter classes to just functions.
